+++ docs/changelog.rst
@@ -4,4 +4,5 @@ Changelog
 2.0.0 (Unreleased)
 ------------------
 
+- #31 Add default contact for ExternalLaboratory and InboundSampleShipment content
 - First version